test: Replace mock local storage implementation with a dedicated mock module
