+++ tests/lib/storage-local.test.ts
@@ -1,26 +1,10 @@
-import { LocalStorageService } from "../../lib/storage-local";
-import { ChatSession } from "../../types/chat";
-
-const mockLocalStorage = (() => {
-  let store: Record<string, string> = {};
-
-  return {
-    getItem: jest.fn((key: string) => store[key] || null),
-    setItem: jest.fn((key: string, value: string) => {
-      store[key] = value.toString();
-    }),
-    removeItem: jest.fn((key: string) => {
-      delete store[key];
-    }),
-    clear: jest.fn(() => {
-      store = {};
-    }),
-  };
-})();
+import { LocalStorageService } from "@/lib/storage-local";
+import { MockStorageLocal } from "@/tests/mocks/storage";
+import { ChatSession } from "@/types/chat";
 
 beforeAll(() => {
   Object.defineProperty(window, "localStorage", {
-    value: mockLocalStorage,
+    value: MockStorageLocal,
   });
 });
